Checking table existence

+++ db/migrate/20130305111144_drop_dropboxparams_table.rb
@@ -1,6 +1,8 @@
 class DropDropboxparamsTable < ActiveRecord::Migration
   def up
-    drop_table :dropboxstrg_params
+    if ActiveRecord::Base.connection.table_exists? 'dropboxstrg_params'
+      drop_table :dropboxstrg_params
+    end
     plugin = Cloudstrg::Cloudstrgplugin.find_by_plugin_name("dropbox")
     if plugin
       plugin.destroy
